+++ quarkc/test/ffi/expected/js/signatures/interfaces/index.js
@@ -14,7 +14,7 @@ exports.A = A;
 
 function A__init_fields__() {}
 A.prototype.__init_fields__ = A__init_fields__;
-A.interfaces_A_ref = quark_ffi_signatures_md.Root.interfaces_A_md;
+A.interfaces_A_ref = null;
 function A_foo() { throw TypeError, '`A.foo` is an abstract method'; }
 A.prototype.foo = A_foo;
 
@@ -29,7 +29,7 @@ exports.B = B;
 
 function B__init_fields__() {}
 B.prototype.__init_fields__ = B__init_fields__;
-B.interfaces_B_ref = quark_ffi_signatures_md.Root.interfaces_B_md;
+B.interfaces_B_ref = null;
 function B_bar() {}
 B.prototype.bar = B_bar;
 
@@ -41,7 +41,7 @@ exports.C = C;
 
 function C__init_fields__() {}
 C.prototype.__init_fields__ = C__init_fields__;
-C.interfaces_C_ref = quark_ffi_signatures_md.Root.interfaces_C_md;
+C.interfaces_C_ref = null;
 function C_foo() {}
 C.prototype.foo = C_foo;
 
@@ -53,7 +53,7 @@ exports.T1 = T1;
 
 function T1__init_fields__() {}
 T1.prototype.__init_fields__ = T1__init_fields__;
-T1.interfaces_T1_ref = quark_ffi_signatures_md.Root.interfaces_T1_md;
+T1.interfaces_T1_ref = null;
 function T1_foo() {}
 T1.prototype.foo = T1_foo;
 
@@ -81,7 +81,7 @@ exports.T2 = T2;
 
 function T2__init_fields__() {}
 T2.prototype.__init_fields__ = T2__init_fields__;
-T2.interfaces_T2_ref = quark_ffi_signatures_md.Root.interfaces_T2_md;
+T2.interfaces_T2_ref = null;
 function T2_foo() {}
 T2.prototype.foo = T2_foo;
 
@@ -109,7 +109,7 @@ exports.T3 = T3;
 
 function T3__init_fields__() {}
 T3.prototype.__init_fields__ = T3__init_fields__;
-T3.interfaces_T3_ref = quark_ffi_signatures_md.Root.interfaces_T3_md;
+T3.interfaces_T3_ref = null;
 function T3_foo() {}
 T3.prototype.foo = T3_foo;
 
@@ -137,7 +137,7 @@ exports.T4 = T4;
 
 function T4__init_fields__() {}
 T4.prototype.__init_fields__ = T4__init_fields__;
-T4.interfaces_T4_ref = quark_ffi_signatures_md.Root.interfaces_T4_md;
+T4.interfaces_T4_ref = null;
 function T4__getClass() {
     return "interfaces.T4";
 }
@@ -165,7 +165,7 @@ exports.T5 = T5;
 
 function T5__init_fields__() {}
 T5.prototype.__init_fields__ = T5__init_fields__;
-T5.interfaces_T5_ref = quark_ffi_signatures_md.Root.interfaces_T5_md;
+T5.interfaces_T5_ref = null;
 function T5_foo() {}
 T5.prototype.foo = T5_foo;
 
@@ -193,8 +193,8 @@ exports.Foo = Foo;
 
 function Foo__init_fields__() {}
 Foo.prototype.__init_fields__ = Foo__init_fields__;
-Foo.interfaces_Foo_ref = quark_ffi_signatures_md.Root.interfaces_Foo_md;
-Foo.quark_List_quark_String__ref = quark_ffi_signatures_md.Root.quark_List_quark_String__md;
+Foo.interfaces_Foo_ref = null;
+Foo.quark_List_quark_String__ref = null;
 function Foo_m1() { throw TypeError, '`Foo.m1` is an abstract method'; }
 Foo.prototype.m1 = Foo_m1;
 
@@ -212,7 +212,7 @@ exports.Bar = Bar;
 
 function Bar__init_fields__() {}
 Bar.prototype.__init_fields__ = Bar__init_fields__;
-Bar.interfaces_Bar_quark_Object__ref = quark_ffi_signatures_md.Root.interfaces_Bar_quark_Object__md;
+Bar.interfaces_Bar_quark_Object__ref = null;
 function Bar_m1() { throw TypeError, '`Bar.m1` is an abstract method'; }
 Bar.prototype.m1 = Bar_m1;
 
@@ -230,7 +230,7 @@ exports.Baz = Baz;
 
 function Baz__init_fields__() {}
 Baz.prototype.__init_fields__ = Baz__init_fields__;
-Baz.interfaces_Baz_ref = quark_ffi_signatures_md.Root.interfaces_Baz_md;
+Baz.interfaces_Baz_ref = null;
 function Baz_m2(arg) {}
 Baz.prototype.m2 = Baz_m2;
 
@@ -261,7 +261,7 @@ exports.RazBar = RazBar;
 
 function RazBar__init_fields__() {}
 RazBar.prototype.__init_fields__ = RazBar__init_fields__;
-RazBar.interfaces_RazBar_ref = quark_ffi_signatures_md.Root.interfaces_RazBar_md;
+RazBar.interfaces_RazBar_ref = null;
 
 // CLASS RazFaz
 function RazFaz() {
@@ -271,7 +271,7 @@ exports.RazFaz = RazFaz;
 
 function RazFaz__init_fields__() {}
 RazFaz.prototype.__init_fields__ = RazFaz__init_fields__;
-RazFaz.interfaces_RazFaz_quark_Object__ref = quark_ffi_signatures_md.Root.interfaces_RazFaz_quark_Object__md;
+RazFaz.interfaces_RazFaz_quark_Object__ref = null;
 
 // CLASS BazBar
 function BazBar() {
@@ -281,7 +281,7 @@ exports.BazBar = BazBar;
 
 function BazBar__init_fields__() {}
 BazBar.prototype.__init_fields__ = BazBar__init_fields__;
-BazBar.interfaces_BazBar_ref = quark_ffi_signatures_md.Root.interfaces_BazBar_md;
+BazBar.interfaces_BazBar_ref = null;
 function BazBar_m1() {}
 BazBar.prototype.m1 = BazBar_m1;
 
@@ -312,7 +312,7 @@ exports.BazFaz = BazFaz;
 
 function BazFaz__init_fields__() {}
 BazFaz.prototype.__init_fields__ = BazFaz__init_fields__;
-BazFaz.interfaces_BazFaz_quark_Object__ref = quark_ffi_signatures_md.Root.interfaces_BazFaz_quark_Object__md;
+BazFaz.interfaces_BazFaz_quark_Object__ref = null;
 function BazFaz_m1() {}
 BazFaz.prototype.m1 = BazFaz_m1;